# pkg/trace: remove heap allocations in signature computations