--- a/releasenotes/notes/trace-agent-no-heap-allocs-in-signature-computations-5418d04185c2e1ca.yaml
+++ b/releasenotes/notes/trace-agent-no-heap-allocs-in-signature-computations-5418d04185c2e1ca.yaml
@@ -1,0 +1,11 @@
+# Each section from every releasenote are combined when the
+# CHANGELOG.rst is rendered. So the text needs to be worded so that
+# it does not depend on any information only available in another
+# section. This may mean repeating some details, but each section
+# must be readable independently of the other.
+#
+# Each section note must be formatted as reStructuredText.
+---
+enhancements:
+  - |
+    APM: improved trace signature computation by avoiding heap allocations.