### What does this PR do?

We compute signatures for all toplevel spans ingested and these used to make some allocations just to copy data that is immediately discarded after the hash computation. This PR copies the code of the hash function and specializes it for our use case to remove the allocations.

Benchmarks for ServiceSignature show that allocs went from 4 and 6 to 0 depending on the hash function and execution time was reduced by 73% and 75% as a result.

 ```
name                     old time/op    new time/op    delta
ServiceSignature_Hash-4     130ns ± 2%      35ns ± 1%   -73.43%  (p=0.008 n=5+5)
ComputeSpanHash-4           180ns ±11%      45ns ± 1%   -75.09%  (p=0.008 n=5+5)

name                     old alloc/op   new alloc/op   delta
ServiceSignature_Hash-4     40.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
ComputeSpanHash-4           48.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)

name                     old allocs/op  new allocs/op  delta
ServiceSignature_Hash-4      4.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
ComputeSpanHash-4            6.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
```

On my memory profiles when benchmarking the agent with single span traces these represented ~20% of the total number of allocations:
<img width="644" alt="image" src="https://user-images.githubusercontent.com/425368/99097570-175bfc00-25d8-11eb-8c17-478b5dec2373.png">


### Motivation

What inspired you to submit this pull request?

### Additional Notes

Anything else we should know when reviewing?

### Describe your test plan

Write there any instructions and details you may have to test your PR.