@gbbr 2020-11-17
If the change is significantly important, we'll need to add [a release note](https://github.com/DataDog/datadog-agent/blob/master/docs/dev/contributing.md#reno), otherwise just add the `changelog/no-changelog` label